DataStructures/BinaryTree/BinaryTreeOrder.test.ts
import BinaryNode from "./BinaryTreeOrder";

import { BTPreOrder } from "./BinaryTreeOrder";
import { BTInOrder } from "./BinaryTreeOrder";
import { BTPostOrder } from "./BinaryTreeOrder";

const tree: BinaryNode<number> = {
    value: 20,
    right: {
        value: 50,
        right: {
            value: 100,
            right: null,
            left: null,
        },
        left: {
            value: 30,
            right: {
                value: 45,
                right: null,
                left: null,
            },
            left: {
                value: 29,
                right: null,
                left: null,
            }
        },
    },
    left: {
        value: 10,
        right: {
            value: 15,
            right: null,
            left: null,
        },
        left: {
            value: 5,
            right: {
                value: 7,
                right: null,
                left: null,
            },
            left: null,
        }
    }
};

test("pre order test", function () {
    expect(BTPreOrder(tree)).toEqual([
        20,
        10,
        5,
        7,
        15,
        50,
        30,
        29,
        45,
        100,
    ]);
});

test("in order test", function () {
    expect(BTInOrder(tree)).toEqual([
        5,
        7,
        10,
        15,
        20,
        29,
        30,
        45,
        50,
        100,
    ]);
});

test("post order test", function () {
    expect(BTPostOrder(tree)).toEqual([
        7,
        5,
        15,
        10,
        29,
        45,
        30,
        100,
        50,
        20,
    ]);
});